backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
qdrant-client
pytest
httpx
